feat(Home): 알림 설정 버튼 위치 이동
