+++ src/css/Home.css
@@ -16,3 +16,18 @@
   width: 30px;
   margin-right: 2px;
 }
+
+#notification-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  position: absolute;
+  top: 10%;
+  right: 10%;
+}
+
+#notification-container p {
+  color: #6f6f6f;
+  font-size: 4px;
+}
